chore: refactor notification component and remove unused imports
